+++ photons_transport/addon.py
@@ -1,3 +1,4 @@
+from photons_transport.session.discovery_options import DiscoveryOptions
 from photons_transport.targets import LanTarget
 
 from photons_app.formatter import MergedOptionStringFormatter
@@ -17,4 +18,7 @@ def __lifx__(collector, *args, **kwargs):
 def __lifx_post__(collector, **kwargs):
     collector.configuration["target_register"].register_type(
         "lan", LanTarget.FieldSpec(formatter=MergedOptionStringFormatter)
+    )
+    collector.register_converters(
+        {"discovery_options": DiscoveryOptions.FieldSpec(formatter=MergedOptionStringFormatter)}
     )