+++ test.sh
@@ -1,2 +1,4 @@
 #!/bin/bash
+unset HARDCODED_DISCOVERY
+unset SERIAL_FILTER
 nosetests --with-noy $@