+++ docs/start/changelog.rst
@@ -16,9 +16,13 @@ ChangeLog
    * Introduced ``collector.run_coro_as_main(coro)`` for running a coroutine as
      the mainline of a program. I also changed the scripts in the examples folder
      to use this method, and cleaned the code in that folder a little.
-   * Another adjustment to shutdown logic to handle async generators better
+   * Another adjustment to shutdown logic to handle shuttind down async
+     generators better
    * Added ``lifx lan:find_ips`` command
    * Fixed the broadcast option to run_with to allow ip addresses
+   * Added discovery options for making photons see only particular devices and/or
+     hard code discovery information for environments where broadcast discovery
+     doesn't work so well. See :ref:`discovery`.
 
 0.22.0 - 21 September 2019
    * Changed the many option on packet definitions to multiple